+++ src/types/channel.ts
@@ -120,6 +120,15 @@ export interface MessageContent {
   replyMessageID?: string;
 }
 
+export interface AllowedMentions {
+  /** An array of allowed mention types to parse from the content. */
+  parse: ("roles" | "users" | "everyone")[];
+  /** Array of role_ids to mention (Max size of 100) */
+  roles?: string[];
+  /** Array of user_ids to mention (Max size of 100) */
+  users?: string[];
+}
+
 export interface GetMessages {
   /** Max number of messages to return(1-100). Defaults to 50. */
   limit?: number;
